Updated newRepo.py
More functionality that may or may not be added to a library that can be used by all of the other functions that we want to have throughout it, I might even publish an entirely new repo just for said library to make git integration a hell whole of a lot easier for most people to do. It seems a lot of kids in my computer science class really struggle with this.